Adding docstring to reset_logging fixture, moving it into the unit test class; moving logging integration tests into the same file
